app/server/requirements.txt
absl-py==0.11.0
astroid==2.5
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
idna==2.10
isort==5.7.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.5.2
Markdown==3.3.3
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydantic==1.7.3
pylint==2.7.1
python-multipart==0.0.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
Werkzeug==1.0.1
wrapt==1.12.1
